added C# OnDestroy

+++ Editor/src/Ouroboros/Scripting/ScriptSystem.cpp
@@ -207,6 +207,10 @@ namespace oo
     {
         if (!isPlaying)
             return;
+        MonoObject* script = scriptDatabase.TryRetrieveObject(uuid, name_space, name);
+        if (script == nullptr)
+            return;
+        ScriptEngine::InvokeFunction(script, "OnDestroy");
         scriptDatabase.Delete(uuid, name_space, name);
     }
     void ScriptSystem::SetScriptEnabled(ScriptDatabase::UUID uuid, const char* name_space, const char* name, bool isEnabled)
@@ -445,6 +449,10 @@ namespace oo
         UUID uuid = e->go->GetInstanceID();
         if (scene.FindWithInstanceID(uuid) == nullptr)
             return;
+        scriptDatabase.ForEachEnabled(uuid, [](MonoObject* object)
+            {
+                ScriptEngine::InvokeFunction(object, "OnDestroy");
+            });
         scriptDatabase.Delete(uuid);
         componentDatabase.Delete(uuid);
     }
